third_party/WebKit/LayoutTests/external/wpt/media-source/mediasource-endofstream-expected.txt

This is a testharness.js-based test.
FAIL MediaSource.endOfStream(): duration truncated to 0 when there are no buffered coded frames assert_equals: expected 0 but got 2
PASS MediaSource.endOfStream(): media element notified that it now has all of the media data
PASS MediaSource.endOfStream(): duration and buffered range end time before and after endOfStream
Harness: the test ran to completion.
